requirement.txt
fastapi==0.101.0
PyJWT==2.7.0
pytest==7.4.0
uvicorn==0.23.2
pydantic==1.10.4
PyJWT==2.7.0
SQLAlchemy=2.0.27
